Create basic div template for each book
The basic structure that each book will be given is and displayed in is
set up. There are two interactive buttons inside them: the trash and
toggle read status buttons. Besides that, each div contains a spot for
book name, author, and number of pages. The page is set up to expand
and create a column a new column for each book so long as there is
enough space.