Make book display look better
Instead of making borders to create separators, new divs are used. This
is done in order to keep the gaps consistent as the div lengthens.
Animations are added to the trash button and status button to make them
scale up a little when hovered. The add new book button is bigger.