Add the inputs of the form
The necessary elements of the form are created, but not styled. There
are inputs for book name, author, and total pages. There is also a
radio button for whether or not the user has read the book. Lastly,
there is an x button which allows the user to close the form.