Make book adding process work properly
As a new book is added, a function runs, which loops through the list
with all the books, makes a card for each book, as well as changing the
information, such as book name and author to what it's supposed to be.
Each card also has a data attribute, which will be useful later in
ensuring both data in the list and the data displayed is matching.